Bugfix: global status randomly fails
In global status command, port occupation will be checked under each
process, however some of them are not ready to listen when the services
just start.

In this case, we have to poll to check when the ports start to listen.
It's not a good way to write a loop everywhere this kind of polling is
used, so we decide to define a try_func function to reduce redundancy.

def try_func(times, func, *args, **kwargs)
"""
times: maximum trials to call the function
func: the function to call
*args: no keyword bounded arguments
**kwargs: keyword bounded arguments
return: function return value or None
"""

If it still fails after *times* calls, this method will raise an
Exception.