Enable qemu version check for certain test
For NVMe option `cmb_size_mb`, we customize test to be run only
on qemu version `>=2.10`. This feature is implemented by a
decorator in infrasim.helper.

Move concrete behavior and attribute of IDEDrvie and SCSIDrive
to their derived class, instead of being in CBaseDrive. Based on
this, we provide NVMeController as a derived class from
CBaseDrive.

**kind of weird**...but we decided to make NVMeController inherits
CBaseDrive instead of CBaseStorageController, due to it needs a
drive backend.

Signed-off-by: bcyan <448162474@qq.com>