Enable share-rw for block device
Added `share-rw` support for block device, which can be parsed
as an attribute of CBaseDrive. It's expected to be a boolean,
which means you need to specify "true|false|on|off" in yml.

Added type protection and unit test on `share-rw`.

Record log and stack for any InfraSimError, so developers won't
have to write one line to log and one line to raise exception.
Now only raise statement is needed.